requirements.txt
azure.cosmos
datetime
python-dateutil
pandas
requests
